# Optimize base and shadow meshes for vertex cache