--- a/scene/resources/3d/importer_mesh.h
+++ b/scene/resources/3d/importer_mesh.h
@@ -114,6 +114,8 @@ public:
 
 	void set_surface_material(int p_surface, const Ref<Material> &p_material);
 
+	void optimize_indices_for_cache();
+
 	void generate_lods(float p_normal_merge_angle, float p_normal_split_angle, Array p_skin_pose_transform_array, bool p_raycast_normals = false);
 
 	void create_shadow_mesh();