@Calinou 2024-07-12
Like `create_shadow_mesh()`, the method in ImporterMesh may be worth exposing to scripting, so that procedural geometry generation scripts can make use of it.