@fire 2024-07-12
At this point, I expect to use both for procedural generation and .. csg, but I am in favour of this.