@mrjustaguy 2024-07-12
https://github.com/godotengine/godot/issues/68959 should be tested with this..