@zeux 2024-07-12
@mrjustaguy That issue should not be affected by this change in isolation for two reasons: 1) this PR only adds the relevant functionality to the glTF import path; adding it to .obj is a matter of adding this to the .obj importer but I'm worried this will cause conflicts with https://github.com/godotengine/godot/pull/94108 so I'd rather do that separately / as part of that change if this ends up getting merged first:

```patch
+       for (int i = 0; i < r_meshes.size(); i++) {
+               r_meshes.get(i)->optimize_indices_for_cache();
+       }
```

2) that issue has a geometry file which is basically a lot of cubes. Due to lack of vertex sharing, in general models with faceted shading - cubes or otherwise - are simultaneously inefficient to render, and mostly not affected by vertex cache optimizations. That said, I would assume #94108 helps somewhat as it adds shadow meshes which should accelerate depth pre-pass/shadow rendering and get a further small boost from this PR.