@mrjustaguy 2024-07-12
I mean that was really a Stress test to compare Godot 3 with Godot 4 primitive performance..

Though I think that there have been a few optimizations relevant to it since I've last looked at it so IDK how Godot 4 compares to 3 Today in that aspect.