@Calinou 2024-07-12
> Would procedural geometry use ImporterMesh or SurfaceTool? Asking because SurfaceTool already exposes `optimize_indices_for_cache`.

Procedural geometry generation is done with SurfaceTool, but ImporterMesh also exposes similar functions so that import scripts can make use of it.